+++ src/mjolnirtools/__init__.py
@@ -1,3 +1,3 @@
 """Convenience tools for users of the Mjolnir HPC cluster."""
 
-__version__ = "1.0.1"
+__version__ = "1.0.3"